+++ Common/src/main/java/ca/lukegrahamlandry/travelstaff/platform/Services.java
@@ -1,7 +1,6 @@
 package ca.lukegrahamlandry.travelstaff.platform;
 
 import ca.lukegrahamlandry.travelstaff.TravelStaffMain;
-import ca.lukegrahamlandry.travelstaff.platform.services.IConfigHelper;
 import ca.lukegrahamlandry.travelstaff.platform.services.INetworkHelper;
 import ca.lukegrahamlandry.travelstaff.platform.services.IPlatformHelper;
 
@@ -11,7 +10,6 @@ public class Services {
 
     public static final IPlatformHelper PLATFORM = load(IPlatformHelper.class);
     public static final INetworkHelper NETWORK = load(INetworkHelper.class);
-    public static final IConfigHelper CONFIG = load(IConfigHelper.class);
 
     public static <T> T load(Class<T> clazz) {