+++ Fabric/src/main/java/ca/lukegrahamlandry/travelstaff/platform/FabricConfigHelper.java
@@ -1,6 +1,0 @@
-package ca.lukegrahamlandry.travelstaff.platform;
-
-import ca.lukegrahamlandry.travelstaff.platform.services.IConfigHelper;
-
-public class FabricConfigHelper implements IConfigHelper {
-}